Changes in build organization
 * targets are now created in top-level build directory
 * more things are now configured through CMake options
 * changed debug build detection from NDEBUG to DEV_BUILD
 * moved po and desktop directories
 * moved last unit test out of src directory

+++ src/graphics/engine/lightman.cpp
@@ -112,7 +112,7 @@ void CLightManager::DebugDumpLights()
             continue;
 
         int deviceLight = -1;
-        for (int j = 0; j < m_lightMap.size(); ++j)
+        for (int j = 0; j < static_cast<int>( m_lightMap.size() ); ++j)
         {
             if (m_lightMap[j] == i)
             {
